Experiments with keyboard bindings for Mac alt keys, done and undone I think.  get_clock fix for Mac

+++ 4ed_font_system_functions.cpp
@@ -297,7 +297,7 @@ system_set_font(System_Functions *system, Partition *part, Render_Font *font, ch
         }
     }
     else{
-        LOGF("font \"%.*s\" name is too long to load in current build (max %u)\n", filename_len, filename, sizeof(font->filename) - 1);
+        LOGF("font \"%.*s\" name is too long to load in current build (max %u)\n", filename_len, filename, (u32)(sizeof(font->filename) - 1));
     }
 }
